# src/api/docs/swagger.yaml
definitions:
  controllers.Component:
    properties:
      image:
        type: string
      name:
        type: string
      network:
        $ref: '#/definitions/controllers.Network'
      resources:
        $ref: '#/definitions/controllers.Resources'
      user_config:
        $ref: '#/definitions/controllers.UserConfig'
    required:
    - name
    type: object
  controllers.ComponentStatus:
    properties:
      name:
        type: string
      state:
        type: string
    type: object
  controllers.GetUserNameResponse:
    properties:
      userID:
        type: string
      userName:
        type: string
    type: object
  controllers.Message:
    properties:
      message:
        type: string
    type: object
  controllers.Network:
    properties:
      expose:
        default: false
        type: boolean
      mac:
        type: string
      port:
        maximum: 65535
        minimum: 0
        type: integer
    type: object
  controllers.Project:
    properties:
      components:
        items:
          $ref: '#/definitions/controllers.Component'
        type: array
      name:
        type: string
      type:
        type: string
      user:
        type: string
    required:
    - components
    - name
    - type
    type: object
  controllers.Resources:
    properties:
      cpu:
        maximum: 2000
        minimum: 0
        type: integer
      memory:
        maximum: 2000
        minimum: 0
        type: integer
    type: object
  controllers.ShortJob:
    properties:
      created:
        type: integer
      id:
        type: string
      status:
        type: string
      type:
        type: string
    type: object
  controllers.UserConfig:
    properties:
      ssh_key:
        type: string
      user:
        type: string
    type: object
  controllers.Utilization:
    properties:
      cpu:
        type: number
      memory:
        type: number
      timestamp:
        type: integer
    type: object
  middleware.AuthSucess:
    properties:
      code:
        type: integer
      expire:
        type: string
      token:
        type: string
    type: object
  middleware.login:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
info:
  contact: {}
paths:
  /health:
    get:
      consumes:
      - application/json
      description: Check if the API is up and running
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Health check
      tags:
      - health
  /login:
    post:
      consumes:
      - application/json
      description: Login
      parameters:
      - description: Login payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/middleware.login'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/middleware.AuthSucess'
      summary: Login
      tags:
      - auth
  /v1/jobs:
    get:
      consumes:
      - application/json
      description: Get all projects from nomad
      parameters:
      - description: Get long project details
        in: query
        name: long
        type: boolean
      - description: Get running projects
        in: query
        name: running
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/controllers.ShortJob'
            type: array
      security:
      - BearerAuth: []
      summary: Get all projects
      tags:
      - job
    post:
      consumes:
      - application/json
      description: Create and submit a job for nomad to deploy
      parameters:
      - description: Project
        in: body
        name: job
        required: true
        schema:
          $ref: '#/definitions/controllers.Project'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.Message'
      security:
      - BearerAuth: []
      summary: Create a project
      tags:
      - job
  /v1/jobs/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a job from nomad
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Purge project
        in: query
        name: purge
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.Message'
      security:
      - BearerAuth: []
      summary: Delete a project
      tags:
      - job
    get:
      consumes:
      - application/json
      description: Get a job from nomad
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get a project
      tags:
      - job
  /v1/jobs/{id}/component/{component}/restart:
    put:
      consumes:
      - application/json
      description: Restart a component in a project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Component name
        in: path
        name: component
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.Message'
      security:
      - BearerAuth: []
      summary: Restart a component
      tags:
      - job
  /v1/jobs/{id}/components:
    get:
      consumes:
      - application/json
      description: Get all components in a project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/controllers.ComponentStatus'
            type: array
      security:
      - BearerAuth: []
      summary: Get all components
      tags:
      - job
  /v1/jobs/{id}/restart:
    put:
      consumes:
      - application/json
      description: Restart a job in nomad
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.Message'
      security:
      - BearerAuth: []
      summary: Restart a project
      tags:
      - job
  /v1/refresh:
    get:
      consumes:
      - application/json
      description: Refresh token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/middleware.AuthSucess'
      security:
      - BearerAuth: []
      summary: Refresh token
      tags:
      - auth
  /v1/resources:
    get:
      consumes:
      - application/json
      description: Get the current CPU and memory usage of all a user's running projects
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.Utilization'
      security:
      - BearerAuth: []
      summary: Get all used resources
      tags:
      - resources
  /v1/resources/{id}:
    get:
      consumes:
      - application/json
      description: Get the current CPU and memory usage of a user's project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.Utilization'
      security:
      - BearerAuth: []
      summary: Get project resource usage
      tags:
      - resources
  /v1/resources/{id}/{component}:
    get:
      consumes:
      - application/json
      description: Get the current CPU and memory usage of a project's component
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Component name
        in: path
        name: component
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.Utilization'
      security:
      - BearerAuth: []
      summary: Get component resource usage
      tags:
      - resources
  /v1/resources/allocated:
    get:
      consumes:
      - application/json
      description: Get the total CPU and memory allocated to a user's running projects
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.Resources'
      security:
      - BearerAuth: []
      summary: Get allocated resources
      tags:
      - resources
  /v1/user:
    get:
      consumes:
      - application/json
      description: Get user name
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.GetUserNameResponse'
      security:
      - BearerAuth: []
      summary: Get user name
      tags:
      - auth
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"